fix: Loc_hal_worker of gnss@2.0-servic crashed
If AGNSS::setServer() is called before
"GnssAPIClient is not ready" meaning FW callbacks are not
yet set by FW. In this case, server configuration
pushed to  mPendingConfig struct. And when FW sets the
callbacks, that is when we try to apply the SUPL server:port
from mPendingConfig. That is when crash happens at
GnssAdapter::setSuplHostServer

Change-Id: Ia34fdd4ff333a6f4cce2bb6ed9081eeefe0aefb2
CRs-fixed: 2869662